Refactors and improves error handling in Spotify client
Renames default controller for clarity and consistency.

Improves error handling in Spotify client by catching exceptions when loading and saving tokens, and during API calls. Adds logging for errors.

Adds contribution guidelines to the README.
