+++ run.py
@@ -1,6 +1,7 @@
 #!/usr/bin/env python3
 import os
 import logging
+
 from app import create_app
